Adds basic support for drawing notes on the score
Currently only draws a rectangle for the duration of the note. No
support for sharps and flats.

This commit also fixes a bug in the midi
player that caused a sleep for a negative number of seconds when
sleeping for multiple frames: "ticks per frame" was a float so "ticks
before next event became negative when "ticks per frame" was repeatedly
subtracted in the frame sleep loop